Rename ToDo, Event, and Deadline
