--- src/main/java/Events.java
@@ -1,23 +1,0 @@
-import java.time.LocalDateTime;
-import java.time.format.DateTimeFormatter;
-
-public class Events extends Task {
-
-    protected LocalDateTime at;
-
-    public Events(String description, String at) {
-        super(description);
-        this.at = LocalDateTime.parse(at, DateTimeFormatter.ofPattern("dd/MM/yyyy HHmm"));
-    }
-
-    @Override
-    public String toString() {
-        return "[E]" + super.toString() + " (at: " +
-                at.format(DateTimeFormatter.ofPattern("MMM d yyyy")) + ")";
-    }
-
-    @Override
-    public String toFileString() {
-        return "E | " + (this.isDone ? 1 : 0) + " | " + this.description + " | " + this.at;
-    }
-}
