--- src/main/java/ToDos.java
@@ -1,15 +1,0 @@
-public class ToDos extends Task {
-    public ToDos(String description) {
-        super(description);
-    }
-
-    @Override
-    public String toString() {
-        return "[T]" + super.toString();
-    }
-
-    @Override
-    public String toFileString() {
-        return "T | " + (this.isDone ? 1 : 0) + " | " + this.description;
-    }
-}
